refactor: move handshake package out of internal
