+++ handshake/domain.go
